(builderjs) Get GitHub Branch Listing
