+++ builderjs/src/builder.ts
@@ -1,0 +1,7 @@
+export const CheckNodeDependencies = async () => {
+  return {};
+};
+
+export const CompileToJson = async () => {
+  return {};
+};
