+++ builderjs/src/tests/index.test.ts
@@ -1,4 +1,4 @@
-import { GetRemoteModulesGithubDirectoryListing } from ".";
+import { GetRemoteModulesGithubDirectoryListing } from "./../index";
 
 async function test() {
   const resp = await GetRemoteModulesGithubDirectoryListing(
