+++ electron-app/src/main.ts
@@ -2,6 +2,8 @@ const { app, BrowserWindow, ipcMain } = require("electron");
 const path = require("path");
 const {
   GetModulesList,
+  CompileToJson,
+  CheckNodeDependencies,
 } = require("/Users/jsb/repos/jsbrittain/phyloflow/builderjs/dist");
 
 const createWindow = () => {
@@ -21,7 +23,7 @@ const createWindow = () => {
   }
 };
 
-async function handleGetRemoteModules(event: any, query: any) {
+async function hBuilderGetRemoteModules(event: any, query: any) {
   const modules = await GetModulesList(query["data"]["content"]["url"]);
   return {
     query: "builder/get-remote-modules",
@@ -29,8 +31,27 @@ async function handleGetRemoteModules(event: any, query: any) {
   };
 }
 
+async function hBuilderCompileToJson(event: any, query: any) {
+  return {
+    query: "builder/compile-to-json",
+    body: {},
+  };
+}
+
+async function hBuilderCheckNodeDependencies(event: any, query: any) {
+  return {
+    query: "builder/compile-to-json",
+    body: {},
+  };
+}
+
 app.whenReady().then(() => {
-  ipcMain.handle("builder/get-remote-modules", handleGetRemoteModules);
+  ipcMain.handle("builder/get-remote-modules", hBuilderGetRemoteModules);
+  ipcMain.handle(
+    "builder/check-node-dependencies",
+    hBuilderCheckNodeDependencies
+  );
+  ipcMain.handle("builder/compile-to-json", hBuilderCompileToJson);
 
   createWindow();
   app.on("activate", () => {
